Made Login and Signup dialog popups
This looks cleaner and keeps the current page while signing in or
registering a new account.
