--- ResearchHub/src/app/+login/login.component.html
@@ -1,18 +1,0 @@
-<md-card>
-  <md-card-title>
-    Login
-  </md-card-title>
-  <span>Email</span>
-  <md-input-container >
-    <input mdInput [(ngModel)]="formUser.username">
-  </md-input-container>
-  <span>Password</span>
-  <md-input-container >
-    <input mdInput [(ngModel)]="formUser.password">
-  </md-input-container>
-  <md-checkbox>Remember Me</md-checkbox>
-
-  <button md-button type="button" (click)="toSignup()">create an account</button>
-  <button md-button type="submit" class="login" (click)="login()">Login</button>
-  
-</md-card>
