--- ResearchHub/src/app/+signup/signup.component.html
@@ -1,21 +1,0 @@
-<md-card>
-  <md-card-title>
-    Sign Up
-  </md-card-title>
-  <span>Email</span>
-  <md-input-container>
-    <input [(ngModel)]="formUser.username" mdInput>
-  </md-input-container>
-  <span>Password</span>
-  <md-input-container>
-    <input [(ngModel)]="formUser.password" mdInput>
-  </md-input-container>
-  <span>Confirm Password</span>
-  <md-input-container>
-    <input mdInput>
-  </md-input-container>
-
-  <button md-button type="button" (click)="toLogin()">go back</button>
-  <button md-button type="submit" class="login" (click)="register()">Sign Up</button>
-
-</md-card>
