+++ ResearchHub/src/app/app.component.ts
@@ -1,6 +1,8 @@
 import { Router } from '@angular/router';
 import { AuthService } from './service/auth.service';
 import { Component } from '@angular/core';
+import { MdDialogConfig, MdDialog } from "@angular/material";
+import { LoginComponent } from "./login/login.component";
 
 @Component({
   selector: 'app-root',
@@ -10,7 +12,7 @@ import { Component } from '@angular/core';
 export class AppComponent {
 
 
-  constructor(private authService: AuthService, private router: Router) {
+  constructor(private authService: AuthService, private router: Router, private dialog: MdDialog) {
 
   }
 
@@ -24,5 +26,12 @@ export class AppComponent {
 
   signOut(): void {
     this.authService.logout();
+  }
+
+  signIn(): void {
+    var dialogConfig = new MdDialogConfig()
+    dialogConfig.height = "230px"
+    dialogConfig.width = "400px"
+    this.dialog.open(LoginComponent, dialogConfig)
   }
 }
