+++ ResearchHub/src/app/signup/signup.component.html
@@ -1,0 +1,18 @@
+<span class="title">Sign Up</span>
+<br />
+<div fxLayout="column" fxLayoutAlign="center stretch">
+  <md-input-container >
+    <input mdInput placeholder="Username" name="Username" [(ngModel)]="formUser.username">
+  </md-input-container>
+  <md-input-container >
+    <input mdInput placeholder="Password" name="password" [(ngModel)]="formUser.password">
+  </md-input-container>
+  <md-input-container >
+    <input mdInput placeholder="Confirm Password" name="confirmPassword" [(ngModel)]="confirmPassword">
+  </md-input-container>
+  <div>    
+    <button md-button type="button" (click)="toLogin()">already have an account?</button>
+    <span fxFlex></span>
+    <button md-button type="submit" class="login" (click)="register()">Sign Up</button>
+  </div>
+</div>
